refactor(zod): rename profileimage.dto -> image.dto with generic types
